requirements_orchestrator.txt
# requirements_orchestrator.txt
fastapi
uvicorn[standard]
httpx
python-multipart
pydantic>=2
python-dotenv